requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
alembic
asyncpg
python-dotenv
authlib
python-jose[cryptography]